# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Pilot API",
    "version": "0.1.2"
  },
  "paths": {
...
    "/api/translation/translate-text": {
      "post": {
        "tags": [
          "translation"
        ],
        "summary": "Post Translate Text",
        "operationId": "post_translate_text_api_translation_translate_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlainTextTranslateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlainTextTranslateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/translation/readme-blocks": {
      "post": {
        "tags": [
          "translation"
        ],
        "summary": "Post Readme Blocks Ephemeral",
        "operationId": "post_readme_blocks_ephemeral_api_translation_readme_blocks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlainTextTranslateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadmeBlocksRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/translation/translate-readme-block": {
      "post": {
        "tags": [
          "translation"
        ],
        "summary": "Post Translate Readme Block Ephemeral",
        "operationId": "post_translate_readme_block_ephemeral_api_translation_translate_readme_block_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadmeBlockTranslateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadmeBlockTranslateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlainTextTranslateRead": {
        "properties": {
          "translated": {
            "type": "string",
            "title": "Translated"
          }
        },
        "type": "object",
        "required": [
          "translated"
        ],
        "title": "PlainTextTranslateRead"
      },
      "PlainTextTranslateRequest": {
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content",
            "description": "待翻译的纯文本"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "PlainTextTranslateRequest"
      },
...
      "ReadmeBlockTranslateRead": {
        "properties": {
          "translated": {
            "type": "string",
            "title": "Translated"
          }
        },
        "type": "object",
        "required": [
          "translated"
        ],
        "title": "ReadmeBlockTranslateRead"
      },
      "ReadmeBlockTranslateRequest": {
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content",
            "description": "待翻译的 Markdown 段落"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ReadmeBlockTranslateRequest"
      },
      "ReadmeBlocksRead": {
        "properties": {
          "blocks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blocks"
          }
        },
        "type": "object",
        "title": "ReadmeBlocksRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}